*: Distinguish between image and image buffer base
FIXME: Split per package
FIXME: Requires mainlining of SMM ECP commands drop

There are distinct concepts of the "image buffer base", which is the
start of the allocated buffer for an image, and the "image base",
which is the start of the image memory. For all image alignments
of at most 4 KB, the two of them are equal. For larger alignments
however, the memory allocated may not be sufficiently aligned and the
image memory must start at a non-zero offset into the image buffer.

The current code does not clearly distinct between the two concepts,
which can cause issues for images with an alignment greater than
4 KB. This patch addresses the issues.

Signed-off-by: Marvin Häuser <mhaeuser@posteo.de>