EmulatorPkg/Host/Unix: Drop dlopen() usage
To assist debugging, i.e. to have the OS load the image debug
symbols, EmulatorPkg/Unix used to load most UEFI executables with OS
APIs, namely dlopen() and dlclose(). After support was added for GDB
and LLDB debugging of images loaded by the EDK II code, this route
became orphaned. Namely, it does not work for a variety of reasons:

1) Support is explicitly disabled for Xcode builds.
2) The dlopen() route is only attempted if the PDB path name ends
with ".pdb", which has not been the case with recent GCC toolchains.
3) By default, the images are not 4K-aligned, which causes some OSes
to deny the image executioon.
4) With recent GCC toolchains, UEFI images are built as ELF
executables, which cannot be dynamically loaded via dlopen().

Drop support for the dlopen() path as realistically it has not been
used or functioning in a long time.

Cc: Andrew Fish <afish@apple.com>
Cc: Ray Ni <ray.ni@intel.com>
Cc: Vitaly Cheptsov <vit9696@protonmail.com>
Signed-off-by: Marvin Häuser <mhaeuser@posteo.de>