MdeModulePkg/CapsuleApp: Add missing EFIAPI specifier
The function CompareFileNameInAlphabet() is passed by pointer to a
generic sorting function. Said pointer declares a function type with
the EFIAPI calling convention. As CompareFileNameInAlphabet() lacks
said specifier, this causes build issues with toolchains that do not
use the EFIAPI calling convention by default, such as CLANG38.

Add the missing specifier to resolve the issue.

Cc: Jian J Wang <jian.j.wang@intel.com>
Cc: Hao A Wu <hao.a.wu@intel.com>
Cc: Liming Gao <gaoliming@byosoft.com.cn>
Cc: Guomin Jiang <guomin.jiang@intel.com>
Cc: Vitaly Cheptsov <vit9696@protonmail.com>
Signed-off-by: Marvin Häuser <mhaeuser@posteo.de>