SecurityPkg/DxeImageVerificationLib: Ignore certs during hash lookup
The current code looks up the image hashes of signed images based on
the hashing algorithms used by the image certificates. This is both
prone to error, and not in-sync with the behaviour of hash enrolment.
SecureBootConfigDxe will only ever enrol image hashes as SHA-256 into
any database, including "db". The UEFI specification makes no
requirements as to which hash algorithms must be used to lookup
hashes in "db".

Align with the SecureBootConfigDxe behaviour to only and always
lookup image hashes as SHA-256 in "db", ignoring the image
certificate layout in the process.

Cc: Jiewen Yao <jiewen.yao@intel.com>
Cc: Jian J Wang <jian.j.wang@intel.com>
Cc: Min Xu <min.m.xu@intel.com>
Cc: Vitaly Cheptsov <vit9696@protonmail.com>
Signed-off-by: Marvin Häuser <mhaeuser@posteo.de>