BaseTools/tools_def: Disable PIE generation for IA32 CLANGDWARF
Commit c25d390 disabled PIE generation for GCC49 and GCC5 as the GCC
toolchains
at the time started enabling it by default. LLVM 15 has now done the same
thing
for all Linux targets [1]. This causes the emission of a GOT with
relocations
that GenFw does not currently support.

As such, disable PIC and PIE for the CLANGDWARF toolchain for IA32
targets, as
already done for GCC49 and GCC5, to support LLVM 15.

[1]
https://github.com/llvm/llvm-project/commit/ca68038d12a23fa7ebb2b1ccbda93c321635e6bf

Cc: Bob Feng <bob.c.feng@intel.com>
Cc: Liming Gao <gaoliming@byosoft.com.cn>
Cc: Yuwei Chen <yuwei.chen@intel.com>
Cc: Pedro Falcato <pedro.falcato@gmail.com>
Signed-off-by: Marvin Häuser <mhaeuser@posteo.de>